Online-Learning-Management-System-LMS-Portal--main/requirements.txt
asgiref==3.9.1
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
djangorestframework==3.16.1
fonttools==4.59.2
fpdf==1.7.2
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.2
packaging==25.0
pandas==2.3.3
pdfconverter==1.0.1
pillow==11.3.0
psycopg2-binary==2.9.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.2
